Web: Files: refactoring: moved onFilesClick to withFileActions HOC
